Moved merge_events from io to events.
